[Fix] Fuzzy search for changed first char (#46)
Fixes incorrect fuzzy search if the first character of the query term was changed as not all initial branches of the trie have been explored. Of course, this requires significantly more work than previously. Hence, this includes a bunch of performance improvements by minimizing allocations and switching to a normal dictionary instead of a sorted one.